set type for s as float64

+++ detectron2/modeling/meta_arch/retinanet.py
@@ -152,8 +152,8 @@ class RetinaNet(nn.Module):
         self.loss_normalizer_momentum = 0.9
 
         if self.training:
-            self.focal_s = torch.nn.Parameter(torch.ones(1, dtype=torch.float32))
-            self.smooth_l1_s = torch.nn.Parameter(torch.ones(1, dtype=torch.float32))
+            self.focal_s = torch.nn.Parameter(torch.ones(1, dtype=torch.float64))
+            self.smooth_l1_s = torch.nn.Parameter(torch.ones(1, dtype=torch.float64))
             self.register_parameter(name="focal_s", param=self.focal_s)
             self.register_parameter(name="smooth_l1_s", param=self.smooth_l1_s)
             np.set_printoptions(threshold=np.inf)
